+++ tests/fake/test_fake_set_maintenance.py
@@ -1,4 +1,5 @@
 import json
+import os
 
 import pytest
 
@@ -10,55 +11,48 @@ from pubtools.pulplib import (
     InvalidDataException,
 )
 
+OWNER = "%s@%s" % (os.environ.get("USER"), os.environ.get("HOSTNAME"))
+
 
 def test_set_maintenance():
 
     controller = FakeController()
 
-    repo1 = Repository(id="repo1")
-    repo2 = Repository(id="repo2")
     maintain_repo = FileRepository(id="redhat-maintenance")
 
-    controller.insert_repository(repo1)
-    controller.insert_repository(repo2)
     controller.insert_repository(maintain_repo)
     # now the maintenance repo is empty
 
     client = controller.client
 
-    regex = "repo1"
+    report = client.get_maintenance_report().result()
 
-    # set maintenance to off doesn't do anything
-    assert client.set_maintenance(regex, enable=False) == None
+    assert isinstance(report, MaintenanceReport)
+    # return an empty report
+    assert report.entries == ()
+    # there's no entries in the report
 
-    # set repo1 to maintenance, a report should be generated and uploaded
-    report = client.set_maintenance(regex)
-
-    report = json.loads(report)
-
-    assert report["last_updated_by"] == "Content Delivery"
-    assert "repo1" in report["repos"].keys()
-    assert "repo2" not in report["repos"].keys()
-    assert report["repos"]["repo1"]["message"] == "Maintenance mode is enabled"
+    report = report.add(repo_ids=["repo1", "repo2"])
+    client.set_maintenance(report).result()
+    # add entries to report and set maintenance
 
     # upload_file and publish should be called once each
     assert len(controller.upload_history) == 1
     assert len(controller.publish_history) == 1
 
     # get_maintenance_report should give a report object
-    report = client.get_maintenance_report()
-    assert isinstance(report, MaintenanceReport)
-    assert report.last_updated_by == "Content Delivery"
+    report = client.get_maintenance_report().result()
+    assert report.last_updated_by == OWNER
+    assert len(report.entries) == 2
 
-    # unset repo1 to maintenance
-    report = client.set_maintenance(regex, enable=False)
-
-    report = json.loads(report)
-    assert report["repos"] == {}
+    report = report.remove(repo_ids=["repo1"], owner="pubtools.pulplib")
+    client.set_maintenance(report).result()
 
     # the report in repo should have updated
-    report = client.get_maintenance_report()
-    assert report.repos == {}
+    report = client.get_maintenance_report().result()
+    assert report.last_updated_by == "pubtools.pulplib"
+    assert len(report.entries) == 1
+    assert report.entries[0].id == "repo2"
 
 
 def test_load_invalid_report_raise_exception():
@@ -66,12 +60,12 @@ def test_load_invalid_report_raise_exception():
 
     client = controller.client
     report = {
-        "last_updated": "2019-08-15T14:21:121Z",
-        "last_updated_by": "Content Delivery",
+        "last_updated": "2019-08-15T14:21:1211",  # invalid timestamp
+        "last_updated_by": "pubtools.pulplib",
         "repos": {
             "repo1": {
                 "message": "Maintenance Mode Enabled",
-                "owner": "Content Delivery",
+                "owner": "pubtools.pulplib",
                 "started": "2019-08-15T14:21:12Z",
             }
         },